+++ packages/utilities/aria-hidden/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @zag-js/aria-hidden
 
+## 0.52.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @zag-js/dom-query@0.52.0
+
 ## 0.51.2
 
 ### Patch Changes